SBERDOMA-62 extract BILLING_PROPERTY_FIELD and BILLING_ACCOUNT_FIELD
