+++ apps/condo/domains/billing/schema/fields.js
@@ -20,8 +20,28 @@ const RAW_DATA_FIELD = {
     isRequired: true,
 }
 
+const BILLING_PROPERTY_FIELD = {
+    schemaDoc: 'Billing property',
+    type: Relationship,
+    ref: 'BillingProperty',
+    isRequired: true,
+    knexOptions: { isNotNullable: true }, // Relationship only!
+    kmigratorOptions: { null: false, on_delete: 'models.CASCADE' },
+}
+
+const BILLING_ACCOUNT_FIELD = {
+    schemaDoc: 'Billing account',
+    type: Relationship,
+    ref: 'BillingAccount',
+    isRequired: true,
+    knexOptions: { isNotNullable: true }, // Relationship only!
+    kmigratorOptions: { null: false, on_delete: 'models.CASCADE' },
+}
+
 module.exports = {
     INTEGRATION_CONTEXT_FIELD,
     IMPORT_ID_FIELD,
     RAW_DATA_FIELD,
+    BILLING_PROPERTY_FIELD,
+    BILLING_ACCOUNT_FIELD,
 }
